test(runtime): group critical YIELD policy regressions
